Split Types.swift
