+++ .gitignore
@@ -4,3 +4,4 @@
 Package.resolved
 
 /.vscode
+/.swiftpm
